Fixed the dynamic navigation layout bug.
Started implementing the "Saved" feature of the app. You can now save entries from the park collection (adding in the other collections next), but it is currently bugged--you can save duplicates of the same entry. To remedy this, add in a condition that checks if the save already exists, and to make things even better, add a Toast.